refactor(backup): extract SQLite snapshot reads into backup/snapshot_db.rs
